Reworked all mods to use new API.
TODO:
Fix getplayerinfo call.
Rework modules loading by main app.

+++ tibiaauto-pub/mods/mod_AddressFinder/AddressFinder.cpp
@@ -214,7 +214,7 @@ long AddressFinder::doComparison()
 		else
 		{
 			TibiaFile.pop_front();
-			TibiaFile.push_back(CMemUtil::GetMemIntValue(++index));
+			TibiaFile.push_back((byte)CMemUtil::GetMemIntValue(++index));
 		}
 	}
 	return foundAddress;
